Initial public release of the code
This is the initial public release of the code and should be runnable as is.  Documentation is still rather sparse and hopefully will be improved before the paper publication.  Note this release does not include our experimental temporal statistics which need further testing and were not included in the paper.